Resolve conflicts with new type crate
