+++ programs/marginfi/src/instructions/marginfi_account/transfer_account.rs
@@ -1,6 +1,6 @@
 use crate::{
     check, check_eq,
-    constants::{ACCOUNT_TRANSFER_FEE, MARGINFI_ACCOUNT_SEED, MOCKS_PROGRAM_ID},
+    constants::{ACCOUNT_TRANSFER_FEE, MOCKS_PROGRAM_ID},
     events::{AccountEventHeader, MarginfiAccountTransferToNewAccount},
     prelude::*,
     state::marginfi_account::MarginfiAccountImpl,
@@ -8,8 +8,11 @@ use crate::{
 use anchor_lang::prelude::*;
 use anchor_lang::solana_program::sysvar::{instructions as ix_sysvar, Sysvar};
 use bytemuck::Zeroable;
-use marginfi_type_crate::types::{
-    LendingAccount, MarginfiAccount, MarginfiGroup, ACCOUNT_DISABLED, ACCOUNT_IN_FLASHLOAN,
+use marginfi_type_crate::{
+    constants::MARGINFI_ACCOUNT_SEED,
+    types::{
+        LendingAccount, MarginfiAccount, MarginfiGroup, ACCOUNT_DISABLED, ACCOUNT_IN_FLASHLOAN,
+    },
 };
 
 pub fn transfer_to_new_account(ctx: Context<TransferToNewAccount>) -> MarginfiResult {
@@ -130,6 +133,11 @@ pub fn transfer_to_new_account_pda(
     anchor_lang::system_program::transfer(ctx.accounts.transfer_fee(), ACCOUNT_TRANSFER_FEE)?;
 
     let mut old_account = ctx.accounts.old_marginfi_account.load_mut()?;
+
+    check!(
+        !old_account.get_flag(ACCOUNT_IN_FLASHLOAN),
+        MarginfiError::AccountInFlashloan
+    );
 
     // Prevent multiple migrations from the same account
     check_eq!(
